Track versions and change how deps are captured
We need to track dependencies per release, so this change adds a key
for releases.  Additionally, since component dependencies are not
static (they can change per release), it's easier for us to track
what dependencies a component has rather than what depends on a
component.  When a component updates, it's easy for us to capture
that component's dependencies rather than trying to figure out all
the different components that require this specific dependency.