RE-1317 Add update validation
The release process needs to be able to recognise what changes have been
made to components in a pull request and whether or not they are valid.
This commit extends the `component` tool to support validating a request
for the release of a new version of a component.

To simply compare the differences between to versions:
component compare --from from-SHA --to to-SHA

To check that the differences constitute only a new version:
component compare --from from-SHA --to to-SHA --verify version