RE-1317 Allow comparison of version on new series
This change allows successful validation of a change that adds a new
version that was part of a new series.

For example:
$ component compare --from from-sha --to to-sha
a-component:
  added:
    releases:
    - series: first
      versions:
      - sha: some-sha
        version: 1.0.0
  deleted: {}

Previously this would fail because the series `first` is being added,
however with this change it succeeds.
$ component compare --from from-sha --to to-sha --verify version
is_product: false
name: a-component
release:
  series: first
  version:
    sha: some-sha
    version: 1.0.0
repo_url: https://github.com/mattt416/a-component